Add a patch for the dvisvgm issue observed in https://github.com/openwebwork/pg/issues/1023.
This patch is applied in the docker build.

It can be applied on your server with the command
`sudo patch -p1 -d / < /opt/webwork/webwork2/docker-config/pgfsys-dvisvmg-bbox-fix.patch.patch`

Note this patch is specific to Ubuntu 22.04, so don't try it on other
versions of Ubuntu or other linux distributions.

This was the change made in https://github.com/pgf-tikz/pgf/pull/1276
that was merged into the master branch of the to the pgf-tikz latex
package on October 23, 2023 that fixes the issue reported in
https://github.com/pgf-tikz/pgf/issues/1275.