Update and fix the docker build.
The dependence on File::Copy::Recursive module was added in #2172, but
not added to the docker build.

Start using node 20 in the docker build.  The node 16 install script
from nodesource is deprecated, and they have imposed a 60 second install
delay when you use it.

Fix the ownership on the courses directory and admin course directory
when the docker entrypoint runs.  Fixing the ownership on the courses
directory and admin course directory will not cause any slow down to the
execution of that script.  It is not a recursive ownership change, just
two directories for which ownership is changed which is super fast.