Add a form to the Accounts Manager for resetting two factor authentication for students.
This form does not allow the user to reset their own two factor
authentication secret, but that of other users at equal or lesser
permission level to their own.

Note that in the admin course if there are multiple admin users, then
one admin user can reset two factor authentication for another.

Also some clean up and issue fixes in the
`htdocs/js/UserList/userlist.js` file with form validation.  The
"change" event handler was being added multiple times to the users list
table. More clean up is needed though (with this and the other pages
with action forms).  There is a lot of redundancy with this
form validation implementation.