Delay file creation when a new file is created in the file manager.
Instead of creating the file when the new file confirmation page is
submitted, create the file when "Save" (or "Save As") is clicked from
the edit page that opens next.

This was requested in issue #2326.

Also, don't focus the `name` input on the edit page.  That is a bit
annoying.  Only focus that input on the confirmation page.  That is the
only page that has a `data` input.