Remove user override check boxes when editing set details for a user.
The dates check on the problem set detail was synchronized with the
check for the user detail page.  Most importantly, a test version's
open, close, and answer dates can not be nullified. In fact, no set's
open, close, and answer dates can be entirely nullified.  In other
words, a global set must have the open, close, and answer dates set.

Note that the set values column was left at the end on this page
(instead of before the user override values as it now is on the user
detail page).  There are issues with moving that first on this page.
There are cases where there are no set values, but there are user values
at all (for example, for the problem seed, problem status, and problem
attempts). So moving set values first would leave an ugly gap between
the label and the user value.

The 10 year limit on future set dates was removed as was done for the
user detail page.

There were some check boxes (not override checks) that had the check box
input inside the label instead of the input and label being siblings.
That is not correct for Bootstrap's expectations for the `form-check`
class.  In particular, the check box and the label "if status is less
than 1" (for the "Randomize Seeds" button) were not aligned correctly
due to this.  So those issues were fixed.