Fill in empty dates with valid dates for date picker groups.
This builds on #2348, and was requested by @Alex-Jordan in that pull
request (see https://github.com/openwebwork/webwork2/pull/2348#pullrequestreview-1918156425).

This makes it so that if a date is not filled in, then it will be filled
by javascript with a date as needed to ensure it satisfies the usual
date requirements on open, reduced scoring, close, and answer dates.

The class value is taken into account for this when editing a set for
users.

This applies on both the problem set detail page and the user detail
page. This javascript is also used on the sets manager page when editing
set dates where this change does not apply, but it is designed to still
work without conflict there.