Switch from $LMSCourseName to $LMSCourseID.
The $LMSCourseName was using the context title from the LMS.
Unfortunately, that may not be unique even within a single LMS.  So
instead this uses the context id.  That must be unique for a single LMS.
That still may not be unique for two different LMS's, but that is not
important here since the identifiers for the external tools from the
different LMS's will be, and those are also checked.