Add the permission level `use_two_factor_auth`.
Roles with this permission level are required to use two factor
authentication to sign in.  Users below this permission level
can sign in directly.  The default user role with this permission is
"student".  But if there is strong opposition to this default, then I
suppose it could be switched to "login_proctor".

Note that even if this is set to "guest", guest users will still be able
to sign in without two factor authentication since it never really makes
sense to have guests (i.e. practice users) use two factor authentication.