Change the fallback date behavior.
Currently if a date is entered into a datepicker input that does not
parse with the format of the current locale (for example if a date is
typed without a time), then a fallback is filled in.  That fallback is
the last displayed valid value if there was one.  If there was an intial
value in the input, then that fallback will initially be the input's
initial value.  It is then updated anytime a new valid value is entered.
However, if there was no initial value for the input, the fallback ends
up being set to the last visible date in what is shown when the calender
popup opens.  That is sommewhat unexpected behavior.

So this changes the way that the fallback works.  If there is an initial
value for the input (as will be the case when editing a set for the
whole class), then that is still used for the initial fallback.
However, if there is no initial value (as will be the case when editing
the set for users), then the class value is used.  In any case, the
fallback will be updated to be the last valid value that is entered
after that.

There is a last case in which the input does not have an initial value
and there also is no class value.  In that case the fallback is the
current date and time.  This usually won't be used on the set detail
page, but applies to the import date shift input on the set list.