Work on separating Utils.pm into smaller Utils/*.pm modules.
Set handling methods are now in WeBWorK::Utils::Sets.  This module
exports the methods:

* format_set_name_internal
* format_set_name_display
* grade_set
* grade_gateway
* grade_all_sets
* is_restricted
* get_test_problem_position
* list_set_versions

Note that `list_set_version` was the only method in
WeBWorK::Utils::Grades which is now deleted.

JITAR problem methods are in WeBWorK::Utils::JITAR.  This module exports
the methods:

* seq_to_jitar_id
* jitar_id_to_seq
* is_jitar_problem_hidden
* is_jitar_problem_closed
* jitar_problem_finished
* jitar_problem_adjusted_status
* prob_id_sort

File system interaction methods are in WeBWorK::Utils::Files.  This
module exports the methods:

* surePathToFile
* readFile
* readDirectory
* listFilesRecursive
* path_is_subdir

Note that the methods createDirectory, makeTempDirectory,
removeTempDirectory that were in WeBWorK::Utils have been deleted, since
they are not used.  The `readFile` method has been updated to use
`slurp` from `Mojo::File`. The `surePathToFile` method now uses
`make_path` from `Mojo::File`.

A bug in the `listFilesRecursive` method was fixed that prevented it
from recursing into symbolic links to directories.  Furthermore, the
method was completely rewritten to use File::Find which should be faster
than the recursive method, and is certainly cleaner codewise.

The `templates/ContentGenerator/Instructor/FileManager/delete.html.ep`
file has been switched from using the the `readDirectory` and
`listFilesRecursive` methods to using the `list` and `list_tree` methods
of `Mojo::File`.  Now that `listFilesRecursive` correctly recurses into
symbolic links, this method reports the incorrect number of files being
deleted in directories containing symbolic links. The actual deletion
does not follow symbolic links, so the files inside symbolic links
should not be counted.  The `list_tree` method does not follow symbolic
links.

With the above change, the only place that `listFilesRecursive` is
actually used is in `templates/ContentGenerator/Instructor/ProblemSetDetail.pm`
to list set and hardcopy headers.  Since `listFilesRecursive` is fixed,
you will now see set headers that are inside symbolic links in a courses
templates directory. Note that this does NOT include set headers that
are in any library listed in `$courseFiles{problibs}` (Library, Contrib,
and capaLibrary by default) since those directories are pruned. That is
true regardless of if those are directories or symbolic links.

Logging methods are provided by WeBWorK::Utils::Logs.  This module
exports the methods:

* writeCourseLog
* writeLog
* writeTimingLogEntry

Note that there is no longer a method `writeCourseLogGivenTime`. Instead
the `writeCourseLog` method accepts an optional $time argument which
defaults to the current time. To be able to do that the accepted
`$message` argument can not be a slurpy parameter.  That was never
actually used anyway though (and would not have worked well if it was).

Date/Time handling methods are provided by WeBWorK::Utils::DateTime.
This module exports the methods:

* before
* between
* after
* formatDateTime
* getDefaultSetDueDate
* parseDateTime
* timeToSec

The constituency_hash, dequote, fisher_yates_shuffle, list2hash,
pretty_print_rh, and ref2string methods in WeBWorK::Utils were deleted
since they are not used.

The rest of the WeBWorK::Utils module is cleaned up.  Signatures are now
used in that file, as well as the new files.

POD has been added for all methods in these files.

Also there is some inclusion clean up.

Also add `follow_skip => 2` to the `File::Find` `find` call in the
`WeBWorK::Utils::Instructor` `getDefList` method.  If two symbolic links
point to the same location, then `find` will die when it is about to
process a file pointed to by the second link because it has already
processed that file. It shouldn't die. Setting `follow_skip => 2` makes
`find` ignore the duplicate. This was discovered while fixing the
`listFilesRecursive` method.