Switch to setting the version explicitly to 0.
Doing this forces the `GD::Barcode::QRcode` package to auto detect the
version, ans works on all platforms to generate a working QR code.
However, on some platforms (well really all of them except the packages
in the Ubuntu repositories) this emits warnings.  To fix that, the
warnings are simply disabled when the image is generated.