Remove lib/WeBWorK/Utils/Task.pm.
The methods from that file are now in lib/WeBWorK/DB/Utils.pm.  The
three methods from lib/WeBWorK/Utils/Task.pm are closely related to some
of the methods already in lib/WeBWorK/DB/Utils.pm, and so they just fit
in there.  Note that the `fake_user` method was deleted, since it is not
used anywhere.

They aren't "tasks" in any case.