Use an unpadded session.
This changes the serialization method used by `Mojo::Sessions` to merely
JSON encode the session without padding.  The default serialization
method JSON encodes the session and then pads it with the letter `Z`
to a length of 1025 characaters.

This results in much smaller session cookies (roughly one fifth the
previous size).  Thus it is possible to be signed in to more courses at
a time without hitting the cookie size limit.

This should perhaps only be considered if the limit on the number of
courses that can be signed into at one time is deemed to be a big enough
problem that something needs to be done.