Remove the bold tag from the unused sub-display error for the admin course.
The errors in the error list that message is added to are html escaped,
and so it can not contain html.  The errors could be prevented from
being escaped, but I see no reason to do so for that one message. None
of the much more important errors added to that list use html.  This
particular message will almost never be seen unless a user is manually
editing the URL with an incorrect subDisplay parameter.