Restructure the problem sets page (the course home page).
The table is replaced with list groups.  There are two or three list
groups, depending on which view you select.  There are two options,
"Show By Date" and "Show By Type".

If "Show By Date" is chosen (this is the default), then sets are grouped
into the three categories "Open Assignments", "Unopen Assigments", and
"Past Due Assignments".  Within the categories the sets are sorted by
urgency (the previous status sort).

If "Show By Type" is chosen, then sets are grouped into the two
categories "Regular Assignments" and "Tests/Quizzes".  Within each group
the sets are ordered by name (the previous name sort).  Note that JITAR
sets are grouped with "Regular Assignments".

Now all sets have an icon before them.  For tests that is the same icon
that was used before.  For other assignments an open book icon is used.
After the icon, the set name is shown with the status below it, and then
other messages (reduced scoring and restricted release messages) below
that.  The download set icon is to the right as before.

Instructors and students now see the same thing regarding the download
set options.  There are no check boxes for an instructor to select and
print multiple sets.  There are still numerous ways that an instructor
can do that.  It is not needed here.  That can be done from the
"Instructor Tools" page or even technically from here by clicking on any
download icon, and then selecting the desired sets on the "Hardcopy
Generator" page.

Several settings are saved in local storage that make it so that when
you return to this page it is the same as it was the last time you
visited it in the same browser.  The last chosen sort method will be
what you see the next time you open the page, and collapsed sections
will still be collapsed.

This is the result of a discussion in our meeting on Wednesday, April 3.
As discussed, some wording may need to be ironed out.

For example, a past due set no longer says "Closed, answers available".
Instead it says "Available for review. Answers available."  There are
too many "availables" in that statement.

Also, is "Regular Assignments" good?  I was going to use "Homework
Assigments", but we are trying to get away from the "homework" name.