Make it so that changing tabs on the course configuration page does not need page reload.
Instead all tabs are present always and are in bootstrap collapses.
Clicking a different tab just toggles visibility.  I.e., the usual nav
tabs setup.

This means that changing tabs does not need a page reload.  The main
advantage of this change is that when settings are saved, all settings
are saved. Even those that are not in the active tab. So you can make
changes in the "General" tab, and without clicking "Save Changes",
select the "Optional Modules" tab and make changes there. Then select
another tab and make changes there.  Then if you click "Save Changes"
all of those changes are saved.

This is much more intuitive behavior, and what most users expect when
seeing a tab layout such as is utilized on this page. As a result, those
new to WeBWorK will often attempt to do what is described in the
paragraph above, and then be surprised to see that the settings from the
not visible tabs are all reset to what they were before the changes the
instructor believed were just changed.

Although this is natural behavior I added a note in the help about this
behavior.  This note is probably more for those that have been using
WeBWorK, and expect the old behavior.

Note that this is built on top of #2392 to avoid needing to resolve
conflicts with that pull request later.