Wait for render completion before saving tmp file when clicking "Check Answers" in the PG editor.
It seems that if the problem rendering request does not complete before
the request to save the temporary file is sent, the streams can get
crossed.  We all know how bad that is!

This seems to only occur when proxying hypnotoad in some way or another.
Somehow hypnotoad manages to keep the streams straight when proxying is
not involved.

Also, don't add the `passwd` and `key` to the form when rendering via
the `render_rpc` endpoint unless cookies are disabled for the request,
and in the case that `session_management_via` is "session_cookie" don't
add the key to the form.

This fixes issue #2394.