Make the PGLM lab problem open via the `render_rpc` endpoint.
Instead of opening the problem using the `editMode` hack of the usual
problem page, render the problem directly in its own page via the
`render_rpc` endpoint.  This means the style hacks in the PGML-lab
problem are also not needed since the elements that were being hidden
are not present in the page to begin with.

The PGMLLab link in a course's templates directory is not needed as the
problem source is read directly from the assets directory, and submitted
to the `render_rpc` endpoint via the `uriEncodedProblemSource` parameter.