Fix an overzealous `->c` addition from #2388.
That pull request changed to passing the controller object to the
ListingDB methods directly instead of using the WebworkWebservice's pass
through methods that make it look like a controller object.  Thus
`$self` needed to be change to `$self->c` in those calls.  However
`$path` was not supposed to be changed to `$path->c`.