Add a "Use score from last check" button to the single problem grader.
If the result score from the last time that "Check Answers" (or "Submit
Answers") is different from the recorded score, then a new button is
shown to the right of the "Problem Score" input (if the percent score is
shown) and another new button is shown to the right of the "Point Value"
input (if the point score is shown).  These buttons read "Use score from
last check: score%" and "Use points from last check: points",
respectively, where the score and points are the result that pg reported
for the last time that "Check Answers" or "Submit Answers" was clicked.
If one of those buttons is clicked, then the last checked score and
points are inserted into the "Point Value" and "Problem Score" inputs.

The idea here is that the instructor may change answers in a student
problem (for example to fix a typo), and then click "Check Answers" to
see the result of those answers.  After doing so the scores from that
check are now shown in the answer part scores above (if there are
multiple parts).  Currently to make those scores take effect in the
problem score that will be saved you need to change one of those part
score inputs away and back.  This just adds a button that will take care
of that for you.  Note that this last checked score must not be put into
the "Point Value" and "Problem Score" inputs, since those must show the
currently recorded score.

I am not sold on the wording shown on the buttons.  Does anyone have a
better suggestion?

Note that the javascript for the single problem grader and the other
grader has been separated into different files.  There is no overlap
between the javascript used by the two anymore, so there is no reason
for this to be in the same file.

Also prevent the default click behavior on 'help-popup's.  This can make
links to '#' scroll to the top of the page.

This pull request is the result of a suggestion by @dlglin.