Add a "Show Correct Answers" button to the problem and test pages.
When this button is used correct answers will be shown with no other
content in the feedback popover, scaffolds will all be open, and the
feedback popovers will open immediately on page load. This is an
instructor only option for now.  I don't think this will be a good idea
to enable for students.

Furthermore, the "Show Correct Answers" buttons on other instructor
pages (pg problem editor, problem library, set detail page, statistics,
and manual problem grader) now use this for showing correct answers.

This works well for most problems, but there are some cases where
answers are tightly grouped particularly when there are on the same line
that the popovers can overlap in not the best way.

This is one potential resolution of https://github.com/openwebwork/pg/issues/1047.